fixed host side test failure for threefryX_64

+++ projects/rocrand/test/internal/test_rocrand_threefry2x64_20_prng.cpp
@@ -65,7 +65,7 @@ using threefry2x64_20_generator_prng_host_tests_types
                                                         ROCRAND_ORDERING_PSEUDO_DEFAULT>>;
 
 using threefry2x64_20_generator_prng_offset_host_tests_types
-    = ::testing::Types<generator_prng_offset_host_tests_params<unsigned int,
+    = ::testing::Types<generator_prng_offset_host_tests_params<unsigned long long,
                                                                threefry2x64_20_generator_host<true>,
                                                                ROCRAND_ORDERING_PSEUDO_DEFAULT>,
                        generator_prng_offset_host_tests_params<float,
